requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
datasets==1.12.1
dill==0.3.4
filelock==3.3.0
fsspec==2021.10.0
huggingface-hub==0.0.19
idna==3.2
importlib-metadata==4.8.1
joblib==1.0.1
multidict==5.2.0
multiprocess==0.70.12.2
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pyarrow==5.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
regex==2021.9.30
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.9.1+cu111
tqdm==4.62.3
transformers==4.11.3
typing-extensions==3.10.0.2
urllib3==1.26.7
xxhash==2.0.2
yarl==1.6.3
zipp==3.6.0
